Load inject as content_script
