+++ src/extension/chrome/manifest.json
@@ -22,6 +22,15 @@
           "file://*/*"],
       "run_at" : "document_start",
       "all_frames" : true
+    },
+    {
+      "js": ["webgpu_inspector.js"],
+      "matches": ["http://*/*",
+        "https://*/*",
+        "file://*/*"],
+      "run_at" : "document_start",
+      "all_frames" : true,
+      "world": "MAIN"
     }
   ],
   "web_accessible_resources": [
